fix ipex 2.3 sdpa attention
