+++ python/llm/src/ipex_llm/transformers/model.py
@@ -52,6 +52,7 @@ from ipex_llm.transformers.gguf.api import load_gguf_model
 
 from .utils import logger, load_state_dict
 from .utils import extract_local_archive_file, get_local_shard_files, load_imatrix_data
+from .patches import patch_flash_attn_import, patch_sdpa_available
 
 patched_training_mode = None
 
@@ -109,19 +110,12 @@ def _load_pre():
     GPTJModel.__init__ = gptj_model_new_init
 
 
-def patch_flash_attn_import(filename: str) -> List[str]:
-    """Work around for https://huggingface.co/microsoft/phi-1_5/discussions/72."""
-    imports = get_imports(filename)
-    if "flash_attn" in imports:
-        imports.remove("flash_attn")
-    return imports
-
-
 class _BaseAutoModelClass:
     HF_MODEL = None
 
     @classmethod
     @patch("transformers.dynamic_module_utils.get_imports", patch_flash_attn_import)
+    @patch("transformers.utils.is_torch_sdpa_available", patch_sdpa_available)
     def from_pretrained(cls,
                         *args,
                         **kwargs):
@@ -549,6 +543,7 @@ class _BaseAutoModelClass:
 
     @classmethod
     @patch("transformers.dynamic_module_utils.get_imports", patch_flash_attn_import)
+    @patch("transformers.utils.is_torch_sdpa_available", patch_sdpa_available)
     def load_low_bit(cls,
                      pretrained_model_name_or_path,
                      *model_args,
